lots of updates see description
this update fixes lots of problems but I may have caused one too: Code blocks are now omitted from the editing functionality, but the editing functionality is acting weird. I condensed the reasons for edits down to simpler terms (the existing ones were a bit necessary and would not fit in some cases; I also added a limited that stops adding reasons into the summary after 200 characters is reached.

I added a namespace container to the editing functionality for numerous advantages in maintenance and extensibility. 

Next on my agenda is to add spell checking (I can actually do this quite well based on a personal test at it that I've tried in the past), and if help is requested on extending the functionality to the inline editing, I'll help with that also.